Harden macOS output routing validation
